citations_masc_computing/requirements.txt
auto_mix_prep==0.2.0
datasets==2.14.5
hydra-core==1.1.2
nltk==3.7
numpy==1.24.4
omegaconf==2.1.2
pandas==2.0.3
pytorch-lightning==1.6.1
torch==1.11.0
transformers==4.18.0
gender_guesser==0.4.0
Genderize==0.3.1
Requests==2.32.3
spacy==2.3.5
fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.8.0/fr_core_news_md-3.8.0.tar.gz
